[sidechain-primitives] move block from root to block.rs
